+++ src/app/layout.tsx
@@ -18,7 +18,7 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
+  title: "AutoResQ",
   description: "Generated by create next app",
 };
